converting to int

+++ commands/utils/set_args.js
@@ -435,9 +435,9 @@ function sync_args_from_cmd(args) {
     }
 
     if ("max_duration" in args) {
-      lt_config["run_settings"]["max_duration"] = parseInt(args["max_duration"]);
+      lt_config["run_settings"]["max_duration"] = parseFloat(args["max_duration"]);
     } else {
-      lt_config["run_settings"]["max_duration"] = parseInt(
+      lt_config["run_settings"]["max_duration"] = parseFloat(
         lt_config["run_settings"]["max_duration"]
       );
     }
